+++ toll_booth/alg_obj/forge/borgs.py
@@ -1,6 +1,6 @@
+from toll_booth.alg_obj.aws.aws_obj.dynamo_driver import DynamoDriver
 from toll_booth.alg_obj.forge.comms.orders import LoadObjectOrder
 from toll_booth.alg_obj.forge.comms.queues import ForgeQueue
-from toll_booth.alg_obj.graph.ogm.ogm import OgmReader
 from toll_booth.alg_obj.graph.ogm.regulators import EdgeRegulator
 
 
@@ -12,7 +12,7 @@ class SevenOfNine:
         self._potential_vertex = metal_order.potential_vertex
         self._rule_entry = metal_order.rule_entry
         self._extracted_data = metal_order.extracted_data
-        self._reader = OgmReader()
+        self._driver = DynamoDriver()
 
     def assimilate(self):
         if self._source_vertex == self._potential_vertex:
@@ -29,7 +29,7 @@ class SevenOfNine:
     def _derive_potential_vertexes(self):
         if self._potential_vertex.is_identifiable:
             return [self._potential_vertex]
-        found_vertexes = self._reader.find_potential_vertexes(self._potential_vertex.object_properties)
+        found_vertexes = self._driver.find_potential_vertexes(self._potential_vertex.object_properties)
         if found_vertexes:
             return found_vertexes
         if self._rule_entry.is_stub:
